A lot of changes on this commit. Still in beta and some more testings should me made, before considering it ok. Basically this adds configurable fields and options using a form. The form it self is a enable / disable. If enable, it will load the configuration loadind the tpl folder. If disable, it runs like the EP4 normal links. One can create customize layouts, they will override the default variables that now are loaded in the preprocess file. If no layout is found, the default config is use ( all fields enable, no reports, no default names ).
The export mode, will look on what fields are enable.
The import, will run with the EP4 beahviour ( if the field is set or not )
The config used is from the last customize layout. ( probably this could be improved)
One can create or delete the layouts ( no edit )
I've try to place all the files in a module folder ( except languages, functions, etc )